Move find batching tests to their own file.
