+++ packages/integration-tests/src/EntityManager.test.ts
@@ -12,9 +12,9 @@ import {
   select,
   update,
 } from "@src/entities/inserts";
-import { aliases, Loaded, sameEntity, setDefaultEntityLimit, setEntityLimit } from "joist-orm";
+import { Loaded, sameEntity, setDefaultEntityLimit, setEntityLimit } from "joist-orm";
 import { Author, Book, Color, newAuthor, newBook, newPublisher, Publisher, PublisherSize } from "./entities";
-import { knex, maybeBeginAndCommit, newEntityManager, numberOfQueries, queries, resetQueryCount } from "./setupDbTests";
+import { knex, maybeBeginAndCommit, newEntityManager, numberOfQueries, resetQueryCount } from "./setupDbTests";
 
 describe("EntityManager", () => {
   it("can load an entity", async () => {
@@ -705,143 +705,6 @@ describe("EntityManager", () => {
       "Cannot set 'publisher' on Author:1 during a flush outside of a entity hook or from afterCommit",
     );
     await flushPromise;
-  });
-
-  it.unlessInMemory("will dedup queries that are loaded at the same time", async () => {
-    await insertPublisher({ name: "p1" });
-    const em = newEntityManager();
-    resetQueryCount();
-    // Given two queries with exactly the same where clause
-    const q1p = em.find(Publisher, { id: "1" });
-    const q2p = em.find(Publisher, { id: "2" });
-    // When they are executed in the same event loop
-    const [q1, q2] = await Promise.all([q1p, q2p]);
-    // Then we issue a single SQL query
-    expect(numberOfQueries).toEqual(1);
-    // And it's the regular/sane query, i.e. not auto-batched
-    expect(queries).toEqual([
-      [
-        `WITH _find (tag, arg0) AS (VALUES ($1::int, $2::int), ($3, $4) )`,
-        ` SELECT array_agg(_find.tag) as _tags, "p".*, p_s0.*, p_s1.*, "p".id as id,`,
-        ` CASE WHEN p_s0.id IS NOT NULL THEN 'LargePublisher' WHEN p_s1.id IS NOT NULL THEN 'SmallPublisher' ELSE 'Publisher' END as __class`,
-        ` FROM publishers as p LEFT OUTER JOIN large_publishers p_s0 ON p.id = p_s0.id`,
-        ` LEFT OUTER JOIN small_publishers p_s1 ON p.id = p_s1.id`,
-        ` JOIN _find ON p.id = _find.arg0 GROUP BY "p".id, p_s0.id, p_s1.id`,
-        ` ORDER BY p.id ASC;`,
-      ].join(""),
-    ]);
-    expect(q1.length).toEqual(1);
-    expect(q2.length).toEqual(0);
-  });
-
-  it.unlessInMemory("will dedup queries with multiple conditions", async () => {
-    await insertAuthor({ first_name: "a1", last_name: "l1" });
-    await insertAuthor({ first_name: "a2", last_name: "l2" });
-    const em = newEntityManager();
-    resetQueryCount();
-    // Given two queries with exactly the same where clause
-    const q1p = em.find(Author, { firstName: "a1", lastName: "l1" });
-    const q2p = em.find(Author, { firstName: "a2", lastName: "l2" });
-    // When they are executed in the same event loop
-    const [q1, a2] = await Promise.all([q1p, q2p]);
-    // Then we issue a single SQL query
-    expect(numberOfQueries).toEqual(1);
-    expect(queries).toEqual([
-      [
-        `WITH _find (tag, arg0, arg1) AS (VALUES`,
-        ` ($1::int, $2::character varying, $3::character varying), ($4, $5, $6) )`,
-        ` SELECT array_agg(_find.tag) as _tags, "a".*`,
-        ` FROM authors as a`,
-        ` JOIN _find ON a.deleted_at IS NULL AND a.first_name = _find.arg0 AND a.last_name = _find.arg1`,
-        ` GROUP BY "a".id`,
-        ` ORDER BY a.id ASC;`,
-      ].join(""),
-    ]);
-  });
-
-  it.unlessInMemory("will dedup queries with complex expressions", async () => {
-    await insertAuthor({ first_name: "a1", last_name: "l1" });
-    await insertAuthor({ first_name: "a2", last_name: "l2" });
-    const em = newEntityManager();
-    resetQueryCount();
-    // Given two queries with exactly the same where clause
-    const [a1, a2] = aliases(Author, Author);
-    const q1p = em.find(Author, { as: a1 }, { conditions: { or: [a1.firstName.eq("a1"), a1.lastName.eq("l1")] } });
-    const q2p = em.find(Author, { as: a2 }, { conditions: { or: [a2.firstName.eq("a2"), a2.lastName.eq("l2")] } });
-    // When they are executed in the same event loop
-    const [q1, q2] = await Promise.all([q1p, q2p]);
-    // Then we issue a single SQL query
-    expect(numberOfQueries).toEqual(1);
-    expect(queries).toEqual([
-      [
-        `WITH _find (tag, arg0, arg1) AS (VALUES`,
-        ` ($1::int, $2::character varying, $3::character varying), ($4, $5, $6) )`,
-        ` SELECT array_agg(_find.tag) as _tags, "a".*`,
-        ` FROM authors as a`,
-        ` JOIN _find ON a.deleted_at IS NULL AND (a.first_name = _find.arg0 OR a.last_name = _find.arg1)`,
-        ` GROUP BY "a".id`,
-        ` ORDER BY a.id ASC;`,
-      ].join(""),
-    ]);
-  });
-
-  it.unlessInMemory("will dedup queries with no conditions", async () => {
-    await insertAuthor({ first_name: "a1" });
-    await insertAuthor({ first_name: "a2" });
-    const em = newEntityManager();
-    resetQueryCount();
-    // Given two queries with exactly the same where clause
-    const q1p = em.find(Author, {});
-    const q2p = em.find(Author, {});
-    // When they are executed in the same event loop
-    const [q1, q2] = await Promise.all([q1p, q2p]);
-    // Then we issue a single SQL query
-    expect(numberOfQueries).toEqual(1);
-    expect(queries).toEqual([
-      `select "a".* from "authors" as "a" where "a"."deleted_at" is null order by "a"."id" asc limit $1`,
-    ]);
-  });
-
-  it.unlessInMemory("does dedup queries with same order bys", async () => {
-    await insertPublisher({ name: "p1" });
-    await insertPublisher({ id: 2, name: "p2" });
-    const em = newEntityManager();
-    resetQueryCount();
-    // Given two queries with exactly the same where clause but different orders
-    const a1p = em.find(Author, { id: "1" }, { orderBy: { firstName: "DESC" }, softDeletes: "include" });
-    const a2p = em.find(Author, { id: "2" }, { orderBy: { firstName: "DESC" }, softDeletes: "include" });
-    // When they are executed in the same event loop
-    const [a1, a2] = await Promise.all([a1p, a2p]);
-    // Then we issue a single SQL query
-    expect(numberOfQueries).toEqual(1);
-    // And it is still auto-batched
-    expect(queries).toMatchInlineSnapshot(`
-      [
-        "WITH _find (tag, arg0) AS (VALUES ($1::int, $2::int), ($3, $4) ) SELECT array_agg(_find.tag) as _tags, "a".* FROM authors as a JOIN _find ON a.id = _find.arg0 GROUP BY "a".id ORDER BY a.first_name DESC;",
-      ]
-    `);
-    // And the results are the expected reverse of each other
-    expect(a1.reverse()).toEqual(a2);
-  });
-
-  it.unlessInMemory("does dedup queries with same order bys via m2os", async () => {
-    const em = newEntityManager();
-    resetQueryCount();
-    // Given two queries with exactly the same where clause but different orders
-    const a1p = em.find(Author, { id: "1" }, { orderBy: { publisher: { id: "ASC" } }, softDeletes: "include" });
-    const a2p = em.find(Author, { id: "2" }, { orderBy: { publisher: { id: "ASC" } }, softDeletes: "include" });
-    // When they are executed in the same event loop
-    const [a1, a2] = await Promise.all([a1p, a2p]);
-    // Then we issue a single SQL query
-    expect(numberOfQueries).toEqual(1);
-    // And it is still auto-batched
-    expect(queries).toMatchInlineSnapshot(`
-      [
-        "WITH _find (tag, arg0) AS (VALUES ($1::int, $2::int), ($3, $4) ) SELECT array_agg(_find.tag) as _tags, "a".* FROM authors as a LEFT OUTER JOIN publishers p ON a.publisher_id = p.id JOIN _find ON a.id = _find.arg0 GROUP BY "a".id, p.id ORDER BY p.id ASC;",
-      ]
-    `);
-    // And the results are the expected reverse of each other
-    expect(a1.reverse()).toEqual(a2);
   });
 
   it("can save tables with self-references", async () => {
